refactor: move depth stack to new module
